requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
Flask==2.2.3
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pinata==0.0.1
pinatapy-vourhey==0.1.8
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
